+++ src/main/browser/log/IRemoteLogger.ts
@@ -1,0 +1,7 @@
+import {ILogEventParams, ILogHandlersModes, LogLevel} from '../../common/log/contracts'
+
+export interface IRemoteLogger {
+	setFileName(value: string)
+
+	send(...logEvents: Array<ILogEventParams<any>>)
+}