+++ HISTORY.md
@@ -1,7 +1,13 @@
 0.1.7 / 2020-09-17
 ===================
 
-  * Change facade: added globalScope and catchUnhandledErrors
+  * added:
+    * globalScope
+    * catchUnhandledErrors
+    * subscribeUnhandledErrors
+    * interceptConsole
+    * interceptEval
+    * catchEvalErrors
   
 0.1.2 - 0.1.6 / 2020-09-17
 ===================